refactor: split InsertSpec into InsertSpec and UpdateSpec
